move Inmemory into an other package
